tests/requirements.txt
pytest==7.2.0
requests==2.28.1
PyJWT==2.6.0
cryptography==38.0.4